+++ src/pages/Index.tsx
@@ -1,7 +1,0 @@
-import { ReportBuilder } from "@/components/ReportBuilder";
-
-const Index = () => {
-  return <ReportBuilder />;
-};
-
-export default Index;